Fix setting of `CPPFLAGS`
The `CPPFLAGS` environment variable is meant for C preprocessor flags,
but was being passed C++ compiler flags. This caused problems when
building with the new compilers as it passed incompatible flags to the C
compiler (e.g. `-std=c++14`). Plus the previous content of `CPPFLAGS`
was being overridden. This commit fixes both issues by appending to
`CPPFLAGS` and adding `CFLAGS` content only.